migrating patient-medications-tabbed.tsx changes to patient-medications.tsx
