+++ src/components/content/medications/patient-medications.tsx
@@ -1,25 +1,21 @@
-import type {
-  FilterChangeEvent,
-  FilterItem,
-} from "@/components/core/filter-bar/filter-bar-types";
 import cx from "classnames";
-import { useEffect, useState } from "react";
 import { MedsHistoryTempProps } from "@/components/content/medications-table-base";
 import {
   BadgeOtherProviderMedCount,
   OtherProviderMedsTable,
   OtherProviderMedsTableProps,
 } from "@/components/content/medications/other-provider-meds-table";
+import { useMedicationFilters } from "@/components/content/medications/patient-medications-filters";
+import { useMedicationSorts } from "@/components/content/medications/patient-medications-sort";
 import { ProviderInactiveMedicationsTable } from "@/components/content/medications/provider-inactive-medications-table";
 import { ProviderMedsTable } from "@/components/content/medications/provider-meds-table";
 import * as CTWBox from "@/components/core/ctw-box";
 import { FilterBar } from "@/components/core/filter-bar/filter-bar";
+import { SortButton } from "@/components/core/sort-button/sort-button";
 import { TabGroup, TabGroupItem } from "@/components/core/tab-group/tab-group";
 import { MedicationStatementModel } from "@/fhir/models";
 import "./patient-medications.scss";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
-import { compact } from "@/utils/nodash";
-import { uniq } from "@/utils/nodash/fp";
 
 export type PatientMedicationsProps = {
   className?: string;
@@ -40,7 +36,6 @@ const tabbedContent = (
 ): TabGroupItem<MedicationStatementModel>[] => [
   {
     key: "medication-list",
-    getPanelClassName: (sm: boolean) => (sm ? "ctw-mt-0" : "ctw-mt-2"),
     display: () => "medication list",
     render: () => (
       <ProviderMedsTable
@@ -51,7 +46,6 @@ const tabbedContent = (
   },
   {
     key: "inactive-provider-records",
-    getPanelClassName: (sm: boolean) => (sm ? "ctw-mt-0" : "ctw-mt-2"),
     display: () => "inactive",
     render: () => (
       <ProviderInactiveMedicationsTable
@@ -78,68 +72,34 @@ export function OtherProviderMedsTableTab({
   onOpenHistoryDrawer,
   onAfterOpenHistoryDrawer,
 }: PatientMedicationsProps) {
-  const [filters, setFilters] = useState<FilterChangeEvent>({});
-  const [records, setRecords] = useState<MedicationStatementModel[]>([]);
-  const { otherProviderMedications, isLoading } =
-    useQueryAllPatientMedications();
+  const { otherProviderMedications } = useQueryAllPatientMedications();
+  const { filters, updateFilters, applyFilters, availableFilters } =
+    useMedicationFilters("other");
+  const { currentSorts, updateSorts, sortOptions, applySorts } =
+    useMedicationSorts();
 
-  useEffect(() => {
-    if (!isLoading && otherProviderMedications) {
-      const filteredRecords = otherProviderMedications.filter((medication) => {
-        if (
-          Array.isArray(filters.providers?.selected) &&
-          filters.providers?.selected.length
-        ) {
-          return filters.providers.selected.includes(
-            medication.lastPrescriber as string
-          );
-        }
-        return true;
-      });
-      setRecords(filteredRecords);
-    }
-  }, [filters, otherProviderMedications, isLoading]);
-
-  const showDismissed = "dismissed" in filters;
-  const showInactive = "inactive" in filters;
-
-  // Create a list of prescriber names to use in a filter
-  const prescriberNames = !otherProviderMedications
-    ? []
-    : (uniq(
-        otherProviderMedications
-          .map((medication) => medication.lastPrescriber)
-          .filter((s) => typeof s === "string")
-      ) as string[]);
-
-  const filterItems: FilterItem[] = compact([
-    {
-      key: "dismissed",
-      type: "tag",
-      icon: "eye",
-      display: ({ active }) =>
-        active ? "dismissed records" : "show dismissed records",
-    },
-    !otherProviderMedications || prescriberNames.length < 2
-      ? null
-      : {
-          key: "providers",
-          type: "checkbox",
-          icon: "clipboard",
-          values: prescriberNames,
-          display: "prescriber",
-        },
-  ]);
-
+  const records = applyFilters(otherProviderMedications ?? []);
   return (
     <>
-      <FilterBar filters={filterItems} handleOnChange={setFilters} />
+      <div className="ctw-flex ctw-flex-wrap ctw-gap-x-2">
+        <SortButton
+          className="ctw-my-2"
+          options={sortOptions}
+          updateSorts={updateSorts}
+          currentSorts={currentSorts}
+        />
+        <FilterBar
+          filters={availableFilters(otherProviderMedications ?? [])}
+          handleOnChange={updateFilters}
+          defaultState={filters.other}
+        />
+      </div>
+
       <OtherProviderMedsTable
-        records={records}
+        records={applySorts(records)}
+        showDismissed={!!filters.other.isArchived}
         handleAddToRecord={handleAddToRecord}
         hideAddToRecord={hideAddToRecord}
-        showDismissed={showDismissed}
-        showInactive={showInactive}
         onOpenHistoryDrawer={onOpenHistoryDrawer}
         onAfterOpenHistoryDrawer={onAfterOpenHistoryDrawer}
       />
